data_gathering/pyproject.toml
[project]
name = "predpovedanie-meksania"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-webhook>=1.4.1",
    "playwright>=1.58.0",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
